requirements.txt
firecrawl-py==2.16.5
pandas>=1.5.0
python-dotenv>=0.19.0
flask>=2.3.0
werkzeug>=2.3.0
